+++ lib/license_finder/license/header_matcher.rb
@@ -2,9 +2,14 @@
 
 module LicenseFinder
   class License
-    HeaderMatcher = Struct.new(:base_matcher) do
+    HeaderMatcher = Struct.new(:base_matcher, :first_n_lines) do
       def matches_text?(text)
-        header = text.split("\n").first || ''
+        n = if first_n_lines.nil?
+              1
+            else
+              first_n_lines
+            end
+        header = text.lines.first(n).join || ''
         base_matcher.matches_text?(header)
       end
     end