+++ spec/lib/license_finder/license/definitions_spec.rb
@@ -82,6 +82,24 @@ describe LicenseFinder::License, 'MIT' do
   end
 end
 
+describe LicenseFinder::License, 'MPL1_1' do
+  subject { described_class.find_by_name 'MPL1_1' }
+
+  it 'should be recognized' do
+    expect(subject).to be
+  end
+
+  describe '#matches_text?' do
+    it "should return true if the text begins with 'Mozilla Public License Version 1.1'" do
+      expect(subject).to be_matches_text 'Mozilla Public License Version 1.1'
+    end
+
+    it "should return false if the text beings with 'Mozilla Public License, version 2.0'" do
+      expect(subject).not_to be_matches_text 'Mozilla Public License, version 2.0'
+    end
+  end
+end
+
 describe LicenseFinder::License, 'MPL2' do
   subject { described_class.find_by_name 'MPL2' }
 
@@ -196,5 +214,12 @@ describe LicenseFinder::License, 'WTFPL' do
     expect(described_class.find_by_name('WTFPL').url).to be
     expect(described_class.find_by_name('WTFPL V2').url).to be
     expect(described_class.find_by_name('Do What The Fuck You Want To Public License').url).to be
+  end
+end
+
+describe LicenseFinder::License, '0BSD' do
+  it 'should be recognized' do
+    expect(described_class.find_by_name('0BSD').url).to be
+    expect(described_class.find_by_name('Zero-Clause BSD').url).to be
   end
 end