+++ spec/lib/license_finder/package_managers/go_modules_spec.rb
@@ -8,64 +8,46 @@ module LicenseFinder
     it_behaves_like 'a PackageManager'
 
     let(:src_path) { '/workspace/code' }
-    let(:sum_path) { "#{src_path}/go.sum" }
+    let(:mod_path) { "#{src_path}/go.mod" }
     let(:vendor_path) { "#{src_path}/vendor" }
+    let(:go_list_format) { '{{ if and (.DepOnly) (not .Standard) }}{{ $mod := (or .Module.Replace .Module) }}{{ $mod.Path }},{{ $mod.Version }},{{ or $mod.Dir .Dir }}{{ end }}' }
     let(:go_list_string) do
       "foo,,/workspace/code/\ngopkg.in/check.v1,v0.0.0-20161208181325-20d25e280405,"\
 "/workspace/LicenseFinder/features/fixtures/go_modules/vendor/gopkg.in/check.v1\n"\
+"gopkg.in/yaml.v2,v2.2.1,/workspace/LicenseFinder/features/fixtures/go_modules/vendor/gopkg.in/yaml.v2\n"\
 'gopkg.in/yaml.v2,v2.2.1,/workspace/LicenseFinder/features/fixtures/go_modules/vendor/gopkg.in/yaml.v2'
     end
-    subject { GoModules.new(project_path: Pathname(src_path), logger: double(:logger, active: nil)) }
+    let(:logger) { double(:logger, active: nil, info: nil) }
+
+    subject { GoModules.new(project_path: Pathname(src_path), logger: logger, log_directory: 'some-directory') }
 
     describe '#current_packages' do
+      let(:go_list_cmd) { "GO111MODULE=on go list -mod=readonly -deps -f '#{go_list_format}' ./..." }
+
       before do
         FakeFS.activate!
-
         FileUtils.mkdir_p(vendor_path)
-        File.write(sum_path, content)
-
-        allow(SharedHelpers::Cmd).to receive(:run).with("GO111MODULE=on go list -m -mod=vendor -f '{{.Path}},{{.Version}},{{.Dir}}' all").and_return(go_list_string)
       end
 
       after do
         FakeFS.deactivate!
       end
 
-      let(:content) do
-        FakeFS.without do
-          fixture_from('go.sum')
-        end
-      end
+      context 'go list is successful' do
+        let(:success_status) { double(Process::Status, success?: true) }
 
-      it 'finds all the packages all go.sum files' do
-        packages = subject.current_packages
-
-        expect(packages.length).to eq 2
-
-        expect(packages.first.name).to eq 'gopkg.in/check.v1'
-        expect(packages.first.version).to eq 'v0.0.0-20161208181325-20d25e280405'
-
-        expect(packages.last.name).to eq 'gopkg.in/yaml.v2'
-        expect(packages.last.version).to eq 'v2.2.1'
-      end
-
-      it 'list packages as Go packages' do
-        packages = subject.current_packages
-
-        expect(packages.first.package_manager).to eq 'Go'
-      end
-
-      context 'when compute is not allowed on vendor' do
         before do
-          allow(SharedHelpers::Cmd).to receive(:run)
-            .with("GO111MODULE=on go list -m -mod=vendor -f '{{.Path}},{{.Version}},{{.Dir}}' all")
-            .and_return(['', "go list -m: can't compute 'all' using the vendor directory\n\t(Use -mod=mod or -mod=readonly to bypass.)\n", 1])
-          allow(SharedHelpers::Cmd).to receive(:run)
-            .with("GO111MODULE=on go list -m -f '{{.Path}},{{.Version}},{{.Dir}}' all")
-            .and_return(go_list_string)
+          File.write(mod_path, content)
+          allow(SharedHelpers::Cmd).to receive(:run).with(go_list_cmd).and_return([go_list_string, nil, success_status])
         end
 
-        it 'finds all the packages all go.sum files' do
+        let(:content) do
+          FakeFS.without do
+            fixture_from('go.mod')
+          end
+        end
+
+        it 'finds all the packages all go.mod files' do
           packages = subject.current_packages
 
           expect(packages.length).to eq 2
@@ -76,12 +58,27 @@ module LicenseFinder
           expect(packages.last.name).to eq 'gopkg.in/yaml.v2'
           expect(packages.last.version).to eq 'v2.2.1'
         end
-      end
-    end
 
-    describe '.prepare_command' do
-      it 'returns the correct package management command' do
-        expect(subject.prepare_command).to eq('GO111MODULE=on go mod tidy && GO111MODULE=on go mod vendor')
+        it 'list packages as Go packages' do
+          packages = subject.current_packages
+
+          expect(packages.first.package_manager).to eq 'Go'
+        end
+      end
+
+      context 'go list failed' do
+        let(:failure_status) { double(Process::Status, success?: false) }
+
+        before do
+          allow(SharedHelpers::Cmd).to receive(:run).with(go_list_cmd).and_return(['', 'some-error-message', failure_status])
+        end
+
+        it 'should print out the error from calling go list' do
+          expect(logger).to receive(:info).with(go_list_cmd, 'did not succeed.', color: :red)
+          expect(logger).to receive(:info).with(go_list_cmd, "Getting the dependencies from go list failed \n\tsome-error-message", color: :red).ordered
+
+          subject.current_packages
+        end
       end
     end